backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
pydantic==1.10.17
python-dotenv==1.0.0
requests==2.31.0
yt-dlp==2023.11.16
google-generativeai==0.3.2
python-multipart==0.0.6
imageio-ffmpeg==0.4.9
assemblyai==0.17.0

